actions/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-firestore
google-cloud-pubsub
numpy
python-dateutil
PyCryptodome
